+++ frontend/package.json
@@ -32,8 +32,8 @@
     "motion": "^12.7.3",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
-    "react-icons": "^5.5.0",
     "react-globe.gl": "^2.33.2",
+    "react-icons": "^5.5.0",
     "react-image": "^4.1.0",
     "react-leaflet": "^5.0.0",
     "react-router-dom": "^6.30.0",
