+++ frontend/src/Components/MainPage/Sidebar.jsx
@@ -14,6 +14,7 @@ import {Link} from "react-router-dom";
 import { motion } from "motion/react";
 import {Img} from "react-image";
 import { cn } from "../../lib/utils";
+import DonationForm from "./NewDonation";
 export function FoodDistributionSidebar() {
   const links = [
     {
@@ -92,6 +93,8 @@ export function FoodDistributionSidebar() {
       console.error("Error getting user data from localStorage:", error);
     }
   }, []);
+
+  
 
   return (
     <Sidebar open={open} setOpen={setOpen}>
